Changes to bullets, enemies, warps can load maps and players can shoot
I made a few changes, especially to constructors. But now we can make warps that move us to a complete new map. Talking of which, now we shouldn't create new objects and entities in the gameplay class, but rather add them to a map and load them from there. Also we can pew pew