added ship and car-key.
when the last boss is killed, the quest is complete and player receives the card key. this can only be used if player is colliding the ship on the asteroid map. if it's used on other circumstances, inventory will not delete the item. if it's used on the ship, victory function is called